requirement.txt
asgiref==3.5.2
bitcoinlib==0.6.5
black==22.6.0
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
colorama==0.4.5
Django==3.2.15
django-mailer==2.2
django-online-users==0.3
django-qr-code==3.1.1
django-tinymce==3.5.0
ecdsa==0.18.0
greenlet==1.1.3
idna==3.3
importlib-metadata==4.12.0
lockfile==0.12.2
mypy-extensions==0.4.3
numpy==1.19.5
pathspec==0.9.0
platformdirs==2.5.2
pycryptodome==3.15.0
pydantic==1.9.2
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.2.1
requests==2.28.1
segno==1.5.2
six==1.16.0
SQLAlchemy==1.4.40
sqlparse==0.4.2
tomli==2.0.1
typed-ast==1.5.4
typing_extensions==4.3.0
urllib3==1.26.11
uuid==1.30
zipp==3.8.1
